requirements.txt
streamlit
google-generativeai
yfinance